requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
pytz==2022.6
sqlparse==0.4.3
psycopg2-binary==2.9.5
django-environ==0.9.0
flake8==5.0.4
isort==5.10.1
django-extensions==3.2.1
django-model-utils==4.2.0
drf-spectacular==0.24.2
drf-spectacular-sidecar==2022.11.1
redis==4.3.4
celery==5.2.7